Move drag handlers to separate file
