+++ edit.js
@@ -14,6 +14,7 @@ import {makeLineMesh, makeTeleportMesh} from './teleport.js';
 
 import targetMeshGeometry from './edit/targetMeshGeometry.js';
 import attachEventListeners from './edit/classlistHandlers.js';
+import dragHandlers from './edit/dragHandlers.js';
 import {
   worldSaveButton, worldRevertButton, micButton, dropdownButton, sandboxButton, newWorldButton,
   packagesSubpage, inventorySubpage, avatarSubpage, avatarSubpageContent,
@@ -1343,64 +1344,6 @@ multiplayerButton.addEventListener('click', async e => {
   worldTools.style.visibility = null;
 }); */
 
-const dropZones = Array.from(document.querySelectorAll('.drop-zone'));
-dropZones.forEach(dropZone => {
-  dropZone.addEventListener('dragenter', e => {
-    dropZone.classList.add('hover');
-  });
-  dropZone.addEventListener('dragleave', e => {
-    dropZone.classList.remove('hover');
-  });
-});
-window.addEventListener('dragend', e => {
-  document.body.classList.remove('dragging-package');
-  dropZones.forEach(dropZone => {
-    dropZone.classList.remove('hover');
-  });
-});
-document.getElementById('inventory-drop-zone').addEventListener('drop', async e => {
-  e.preventDefault();
-
-  const jsonItem = Array.from(e.dataTransfer.items).find(i => i.type === 'application/json+package');
-  if (jsonItem) {
-    const s = await new Promise((resolve, reject) => {
-      jsonItem.getAsString(resolve);
-    });
-    const j = JSON.parse(s);
-    let {name, dataHash, id, iconHash} = j;
-    if (!dataHash) {
-      const p = pe.children.find(p => p.id === id);
-      dataHash = await p.getHash();
-    }
-
-    const inventory = loginManager.getInventory();
-    inventory.push({
-      name,
-      dataHash,
-      iconHash,
-    });
-    await loginManager.setInventory(inventory);
-  }
-});
-document.getElementById('avatar-drop-zone').addEventListener('drop', async e => {
-  e.preventDefault();
-
-  const jsonItem = Array.from(e.dataTransfer.items).find(i => i.type === 'application/json+package');
-  if (jsonItem) {
-    const s = await new Promise((resolve, reject) => {
-      jsonItem.getAsString(resolve);
-    });
-    const j = JSON.parse(s);
-    let {dataHash, id} = j;
-    if (!dataHash) {
-      const p = pe.children.find(p => p.id === id);
-      dataHash = await p.getHash();
-    }
-
-    await loginManager.setAvatar(dataHash);
-  }
-});
-
 window.addEventListener('avatarchange', e => {
   const p = e.data;
 
@@ -2143,3 +2086,4 @@ window.addEventListener('popstate', e => {
 _handleUrl(window.location.href);
 
 attachEventListeners();
+dragHandlers(pe, loginManager);
